Rename node and primitive files
