+++ Specs/Scene/ModelExperimental/ModelExperimentalRuntimeNodeSpec.js
@@ -5,13 +5,13 @@ import {
   InstancingPipelineStage,
   Matrix3,
   Matrix4,
-  ModelExperimentalNode,
+  ModelExperimentalRuntimeNode,
   ModelMatrixUpdateStage,
   NodeStatisticsPipelineStage,
   Quaternion,
 } from "../../../Source/Cesium.js";
 
-describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
+describe("Scene/ModelExperimental/ModelExperimentalRuntimeNode", function () {
   const mockNode = {
     matrix: Matrix4.IDENTITY,
   };
@@ -58,7 +58,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
 
   it("throws for undefined node", function () {
     expect(function () {
-      return new ModelExperimentalNode({
+      return new ModelExperimentalRuntimeNode({
         node: undefined,
         transform: transform,
         transformToRoot: transformToRoot,
@@ -70,7 +70,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
 
   it("throws for undefined transform", function () {
     expect(function () {
-      return new ModelExperimentalNode({
+      return new ModelExperimentalRuntimeNode({
         node: mockNode,
         transform: undefined,
         transformToRoot: transformToRoot,
@@ -82,7 +82,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
 
   it("throws for undefined transform to root", function () {
     expect(function () {
-      return new ModelExperimentalNode({
+      return new ModelExperimentalRuntimeNode({
         node: mockNode,
         transform: transform,
         transformToRoot: undefined,
@@ -94,7 +94,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
 
   it("throws for undefined scene graph", function () {
     expect(function () {
-      return new ModelExperimentalNode({
+      return new ModelExperimentalRuntimeNode({
         node: mockNode,
         transform: transform,
         transformToRoot: transformToRoot,
@@ -106,7 +106,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
 
   it("throws for undefined children", function () {
     expect(function () {
-      return new ModelExperimentalNode({
+      return new ModelExperimentalRuntimeNode({
         node: mockNode,
         transform: transform,
         sceneGraph: mockSceneGraph,
@@ -117,7 +117,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("constructs", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -146,7 +146,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   it("constructs with default transform parameters", function () {
     const mockNodeWithNoMatrix = {};
 
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNodeWithNoMatrix,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -172,7 +172,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
       scale: new Cartesian3(1.0, 1.0, 2.0),
     };
 
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNodeWithParameters,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -192,7 +192,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("setting translation throws if node was constructed with matrix", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -206,7 +206,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("setting rotation throws if node was constructed with matrix", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -220,7 +220,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("setting scale throws if node was constructed with matrix", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -234,7 +234,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("setting morphWeights throws if given different length array", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -254,7 +254,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
       translation: Cartesian3.ZERO,
     };
 
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNodeWithTranslation,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -283,7 +283,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
       rotation: Quaternion.IDENTITY,
     };
 
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNodeWithRotation,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -313,7 +313,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
       scale: new Cartesian3(1.0, 1.0, 1.0),
     };
 
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNodeWithScale,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -339,7 +339,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
       morphWeights: [0.0, 0.0, 0.0],
     };
 
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNodeWithWeights,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -363,7 +363,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
         attributes: [],
       },
     };
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: instancedMockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -387,7 +387,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("getChild throws for undefined index", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -401,7 +401,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("getChild throws for invalid index", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -415,7 +415,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("getChild throws for out of range index", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -432,7 +432,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("getChild works", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -446,7 +446,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("sets transform without replacing original", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
@@ -468,7 +468,7 @@ describe("Scene/ModelExperimental/ModelExperimentalNode", function () {
   });
 
   it("updateComputedTransform updates computedTransform matrix", function () {
-    const node = new ModelExperimentalNode({
+    const node = new ModelExperimentalRuntimeNode({
       node: mockNode,
       transform: transform,
       transformToRoot: transformToRoot,
